Add files via upload
This project is developed as a project for the Web Application Development subject, as a course requirement of Bachelor of Computer Science (Software Development) with Honours, at Universiti Teknikal Malaysia Melaka.

This system is about non-biodegradable waste management. The system enables staff to record the waste collected as transactions, and the system can show the current capacity of the collection bins.

The frontend of the system is written with HTML and the Bootstrap framework is used as the CSS. The backend of the system is written with PHP, and database with MySQL.

This system is still a prototype and development still in progress.